consolidate all the semver code
